+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-28T09:18:44.796Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T10:29:45.432Z")
     r"""The date and time the resource was last updated"""