+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-28T09:18:44.419Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T10:29:45.003Z")
     r"""The date and time the resource was last updated"""