+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-28T09:18:48.117Z")
+    created: Optional[datetime] = parse_datetime("2025-10-28T10:29:48.802Z")
     r"""The date and time the resource was created"""