+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-28T09:18:44.796Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T10:29:45.432Z")
     r"""The date and time the resource was last updated"""